Fix #211: avoid wget which touches temp file
This is a follow up of https://github.com/lampepfl/bench/pull/170,
where some the network file system creates instability.
From manua testing on the server, `wget` reles on temp files,
which takes more than 1min to save the file and the content
of the file is incorrect.

Using of `curl` avoids the problem.